Project 014 - Image Generator/requirements.txt
streamlit
diffusers
torch
torchvision
Pillow
transformers
accelerate